combines all resume data entry d2d tests in a single file
